refact: Login element grouped to component
